backend/requirements.txt
fastapi
uvicorn
pandas
numpy
sqlalchemy
pydantic
python-dotenv
scikit-learn
openai
faiss-cpu